requirements.txt
bcrypt==4.1.3
boto3==1.34.134
botocore==1.34.134
cffi==1.16.0
click==8.1.7
cryptography==42.0.8
jmespath==1.0.1
paramiko==3.4.0
pycparser==2.22
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.2
scp==0.15.0
six==1.16.0
urllib3==2.2.2
